Add LSystem class
The class currently has two properties: `axiom`, which is the value of
the LSystem at level 0, and `rules`, which is an object that maps each
letter in the LSystem's alphabet to a valid word in the system.
Currently no validation is done.

The class also has a `charsAtLevel()` generator, which lazily provides
the characters of the string generated by applying the rules to each
character of the axiom some given number of times. Because the length of
the string at a given level generally grows exponentially with the
level, we use a stack-based traversal that guarantees that no more than
`(M - 1) * level` entries are stored in memory at any given time, where
`M` is the length of the longest string value in the `rules` map, and
`level` is the desired level. For example, if the longest value in the
rules map is 3 and we're generating the 50th level of the system, there
will be no more than (3 - 1) * 50 = 100 entries in the stack at any
given time.